+++ python/lsst/ts/ess/csc/data_client/controller_data_client.py
@@ -295,11 +295,8 @@ class ControllerDataClient(common.BaseDataClient):
                         f"Encountered at least {MAX_ALLOWED_READ_TIMEOUTS} timeouts. Raising error."
                     )
                     raise
-            except RuntimeError as e:
-                self.log.exception(f"read_loop failed: {e}")
-                raise
-            except Exception:
-                self.log.exception("read_loop failed")
+            except Exception as e:
+                self.log.exception(f"read_loop failed: {e!r}")
                 raise
 
     async def read(self) -> dict:
@@ -427,28 +424,34 @@ class ControllerDataClient(common.BaseDataClient):
         RuntimeError
             If the response code is common.ResponseCode.DEVICE_READ_ERROR
         """
-        device_configuration = self.device_configurations.get(sensor_name)
-        if device_configuration is None:
-            raise RuntimeError(f"No device configuration for sensor_name={sensor_name}")
-        if response_code == common.ResponseCode.OK:
-            telemetry_method = self.telemetry_dispatch_dict.get(
-                device_configuration.sens_type
-            )
-            if telemetry_method is None:
+        try:
+            device_configuration = self.device_configurations.get(sensor_name)
+            if device_configuration is None:
                 raise RuntimeError(
-                    f"Unsupported sensor type {device_configuration.sens_type}"
+                    f"No device configuration for sensor_name={sensor_name}"
                 )
-            await telemetry_method(
-                sensor_name=sensor_name,
-                timestamp=timestamp,
-                response_code=response_code,
-                sensor_data=sensor_data,
-            )
-        elif response_code == common.ResponseCode.DEVICE_READ_ERROR:
-            raise RuntimeError(
-                f"Error reading sensor {sensor_name}. Please check the hardware."
-            )
-        else:
-            self.log.warning(
-                f"Ignoring telemetry for sensor {sensor_name} with unknown response code {response_code}"
-            )
+            if response_code == common.ResponseCode.OK:
+                telemetry_method = self.telemetry_dispatch_dict.get(
+                    device_configuration.sens_type
+                )
+                if telemetry_method is None:
+                    raise RuntimeError(
+                        f"Unsupported sensor type {device_configuration.sens_type}"
+                    )
+                await telemetry_method(
+                    sensor_name=sensor_name,
+                    timestamp=timestamp,
+                    response_code=response_code,
+                    sensor_data=sensor_data,
+                )
+            elif response_code == common.ResponseCode.DEVICE_READ_ERROR:
+                raise RuntimeError(
+                    f"Error reading sensor {sensor_name}. Please check the hardware."
+                )
+            else:
+                self.log.warning(
+                    f"Ignoring telemetry for sensor {sensor_name} with unknown response code {response_code}"
+                )
+        except Exception as e:
+            self.log.exception(f"process_telemetry failed: {e!r}")
+            raise